back to floats

+++ src/bpa/primitives.h
@@ -5,9 +5,9 @@
 #include <list>
 
 struct Vertex {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
 typedef size_t VertexIndex;
@@ -26,9 +26,9 @@ struct Triangle {
 };
 
 struct Vector {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
 typedef std::vector<Vertex> Vertices;
@@ -42,15 +42,15 @@ bool operator==(const Triangle triangle1, const Triangle triangle2);
 bool operator!=(const Triangle triangle1, const Triangle triangle2);
 
 Vector conn(const Vertex from, const Vertex to);
-double dist(const Vertex from, const Vertex to);
+float dist(const Vertex from, const Vertex to);
 bool equals(const Vertex vertex1, const Vertex vertex2);
 bool same(const Vertex vertex1, const Vertex vertex2);
 
 Vertex toVertex(const Vector vector);
 Vector operator+(const Vector vector1, const Vector vector2);
 Vector operator-(const Vector vector1, const Vector vector2);
-double operator*(const Vector vector1, const Vector vector2);
-Vector operator*(const double scale, const Vector vector);
+float operator*(const Vector vector1, const Vector vector2);
+Vector operator*(const float scale, const Vector vector);
 Vector cross(const Vector vector1, const Vector vector2);
-double len(const Vector vector);
-Vector setMag(const Vector vector, const double newMag);
+float len(const Vector vector);
+Vector setMag(const Vector vector, const float newMag);
